+++ src/Calculator/Views/MainPage.xaml.cs
@@ -15,6 +15,8 @@ using Windows.UI.Core;
 using Windows.UI.ViewManagement;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Automation;
+using Windows.UI.Xaml.Automation.Peers;
+using Windows.UI.Xaml.Automation.Provider;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Controls.Primitives;
 using Windows.UI.Xaml.Data;
@@ -38,6 +40,9 @@ namespace CalculatorApp
             Application.Current.Suspending += App_Suspending;
             m_model.PropertyChanged += OnAppPropertyChanged;
             m_accessibilitySettings = new AccessibilitySettings();
+
+            Loaded += OnLoaded;
+            Unloaded += OnUnloaded;
 
             if(Utilities.GetIntegratedDisplaySize(out var sizeInInches))
             {
@@ -147,6 +152,17 @@ namespace CalculatorApp
             }
 
             m_model.Initialize(initialMode);
+        }
+
+        private void OnLoaded(object sender, RoutedEventArgs e)
+        {
+            // Register the system back requested event
+            SystemNavigationManager.GetForCurrentView().BackRequested += System_BackRequested;
+        }
+
+        private void OnUnloaded(object sender, RoutedEventArgs e)
+        {
+            SystemNavigationManager.GetForCurrentView().BackRequested -= System_BackRequested;
         }
 
         private void UpdatePopupSize(Windows.UI.Core.WindowSizeChangedEventArgs e)
@@ -578,19 +594,31 @@ namespace CalculatorApp
             NarratorNotifier.Announce(announcement);
         }
 
-        private void TitleBarBackButtonClick(object sender, RoutedEventArgs e)
+        private bool ShouldShowBackButton(bool isAlwaysOnTop, bool isPopupOpen)
         {
-            CloseSettingsPopup();
-        }
-
-        private Visibility ShouldShowBackButton(bool isAlwaysOnTop, bool isPopupOpen)
-        {
-            return !isAlwaysOnTop && isPopupOpen ? Visibility.Visible : Visibility.Collapsed;
+            return !isAlwaysOnTop && isPopupOpen;
         }
 
         private double NavigationViewOpenPaneLength(bool isAlwaysOnTop)
         {
             return isAlwaysOnTop ? 0 : (double)Application.Current.Resources["SplitViewOpenPaneLength"];
+        }
+
+        private void System_BackRequested(object sender, BackRequestedEventArgs e)
+        {
+            if (!e.Handled && BackButton.IsEnabled)
+            {
+                var buttonPeer = new ButtonAutomationPeer(BackButton);
+                IInvokeProvider invokeProvider = buttonPeer.GetPattern(PatternInterface.Invoke) as IInvokeProvider;
+                invokeProvider.Invoke();
+
+                e.Handled = true;
+            }
+        }
+
+        private void BackButton_Click(object sender, RoutedEventArgs e)
+        {
+            CloseSettingsPopup();
         }
 
         private CalculatorApp.Calculator m_calculator;
